+++ src/rstrings.jl
@@ -19,7 +19,7 @@ function sadd(value::AbstractString, ttl::AbstractString)
     if isa(value_n, Nothing)
         value_n = value
     end
-    ttl_p = tryparse(Int128, ttl)
+    ttl_p = tryparse(Int, ttl)
     if isa(ttl_p, Nothing)
         println("ttl not a valid integer - got '$ttl' tt forced to nothing")
     end
